fixed random scalars
